requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
Pillow>=10.0.0
opencv-python-headless>=4.9.0
scikit-learn>=1.4.0
scipy>=1.12.0
numpy>=1.26.0
reportlab>=4.1.0
python-multipart>=0.0.9
pytest>=8.0.0
httpx>=0.27.0
